xe: gemm: jit: widen the preassigned token map key
8-bit is not enough for grf512 kernels.

+++ src/gpu/intel/gemm/jit/generator/pieces/state.hpp
@@ -145,7 +145,7 @@ struct CommonState {
     std::array<VirtualFlag, 8> activeVFlags;
     VirtualFlagAllocator raVFlag;
     TokenAllocator tokenAllocator;
-    std::vector<std::pair<uint8_t, int8_t>> tokenMap;
+    std::vector<std::pair<uint16_t, int8_t>> tokenMap;
     ngen::Subregister readFailures;
     ngen::Subregister fusedID;
     ngen::Subregister lsDescConstant[4];
